カメラnew element

+++ db/migrate/20191231232320_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.1]
     create_table :items do |t|
       t.string :name
       t.integer :price
-      t.integer :jan
+      t.bigint :jan
       t.integer :baika
       t.integer :genka
       t.integer :stock,default: 0
